WIP: fix schema workers overwriting each others work when they both use the same file.
This introduces a new bug where schemas with multiple definitions (e.g. pets.json, companions.json, election.json) merge the root anyOf arrays for when `success` is true or false.
